+++ src/Schema/Factories/FieldFactory.php
@@ -61,8 +61,8 @@ class FieldFactory
         $fieldDefinitionNode = $fieldValue->getField();
 
         // Directives have the first priority for defining a resolver for a field
-        /** @var \Nuwave\Lighthouse\Support\Contracts\FieldResolver $resolverDirective */
-        if ($resolverDirective = $this->directiveFactory->exclusiveOfType($fieldDefinitionNode, FieldResolver::class)) {
+        $resolverDirective = $this->directiveFactory->exclusiveOfType($fieldDefinitionNode, FieldResolver::class);
+        if ($resolverDirective instanceof FieldResolver) {
             $fieldValue = $resolverDirective->resolveField($fieldValue);
         } else {
             $fieldValue = $fieldValue->useDefaultResolver();
@@ -80,7 +80,11 @@ class FieldFactory
             ->send($fieldValue)
             ->through($fieldMiddleware->all())
             ->via('handleField')
-            ->thenReturn()
+            // TODO replace when we cut support for Laravel 5.6
+            //->thenReturn()
+            ->then(static function (FieldValue $fieldValue): FieldValue {
+                return $fieldValue;
+            })
             ->getResolver();
 
         $fieldValue->setResolver(