+++ src/Subscriptions/Serializer.php
@@ -66,19 +66,7 @@ class Serializer implements ContextSerializer
 
         $request->setUserResolver(
             function () use ($rawUser) {
-                $user = $this->getRestoredPropertyValue($rawUser);
-
-                // This is here for backwards compatibility, before the Laravel
-                // `SerializesAndRestoresModelIdentifiers` trait was used to
-                // serialize the user data it was `serialize`d separately
-                // This allows existing subscriptions to be seamlessly
-                // upgraded without breaking an active subscription
-                // TODO remove this fallback in v5
-                if (is_string($user)) {
-                    return unserialize($rawUser);
-                }
-
-                return $user;
+                return $this->getRestoredPropertyValue($rawUser);
             }
         );