+++ src/Subscriptions/Subscriber.php
@@ -7,15 +7,12 @@ use GraphQL\Language\AST\NodeList;
 use GraphQL\Type\Definition\ResolveInfo;
 use GraphQL\Utils\AST;
 use Illuminate\Support\Str;
-use Nuwave\Lighthouse\Exceptions\SubscriptionException;
 use Nuwave\Lighthouse\Subscriptions\Contracts\ContextSerializer;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use Serializable;
 
 class Subscriber implements Serializable
 {
-    public const MISSING_OPERATION_NAME = 'Must pass an operation name when using a subscription.';
-
     /**
      * A unique key for the subscriber.
      *
@@ -38,14 +35,14 @@ class Subscriber implements Serializable
     public $query;
 
     /**
-     * The name of the queried operation.
+     * The name of the queried field.
      *
-     * TODO replace with subscription field name, is guaranteed be be unique because of
+     * Guaranteed be be unique because of
      * @see \GraphQL\Validator\Rules\SingleFieldSubscription
      *
      * @var string
      */
-    public $operationName;
+    public $fieldName;
 
     /**
      * The root element of the query.
@@ -70,31 +67,23 @@ class Subscriber implements Serializable
 
     /**
      * @param  array<string, mixed>  $args
-     *
-     * @throws \Nuwave\Lighthouse\Exceptions\SubscriptionException
      */
     public function __construct(
         array $args,
         GraphQLContext $context,
         ResolveInfo $resolveInfo
     ) {
-        $operation = $resolveInfo->operation;
-        // TODO remove that check and associated tests once graphql-php covers that validation https://github.com/webonyx/graphql-php/pull/644
-        if ($operation === null) {
-            throw new SubscriptionException(self::MISSING_OPERATION_NAME);
-        }
-
-        $operationName = $operation->name;
-
-        // TODO remove that check and associated tests once graphql-php covers that validation https://github.com/webonyx/graphql-php/pull/644
-        if (! $operationName) { // @phpstan-ignore-line TODO remove when upgrading graphql-php
-            throw new SubscriptionException(self::MISSING_OPERATION_NAME);
-        }
-        $this->operationName = $operationName->value;
-
+        $this->fieldName = $resolveInfo->fieldName;
         $this->channel = self::uniqueChannelName();
         $this->args = $args;
         $this->context = $context;
+
+        /**
+         * Must be here, since webonyx/graphql-php validated the subscription.
+         *
+         * @var \GraphQL\Language\AST\OperationDefinitionNode $operation
+         */
+        $operation = $resolveInfo->operation;
 
         $this->query = new DocumentNode([
             'definitions' => new NodeList(array_merge(
@@ -111,14 +100,14 @@ class Subscriber implements Serializable
      */
     public function unserialize($subscription): void
     {
-        $data = json_decode($subscription, true);
+        $data = \Safe\json_decode($subscription, true);
 
         $this->channel = $data['channel'];
         $this->topic = $data['topic'];
         $this->query = AST::fromArray( // @phpstan-ignore-line We know this will be exactly a DocumentNode and nothing else
             unserialize($data['query'])
         );
-        $this->operationName = $data['operation_name'];
+        $this->fieldName = $data['field_name'];
         $this->args = $data['args'];
         $this->context = $this->contextSerializer()->unserialize(
             $data['context']
@@ -130,24 +119,16 @@ class Subscriber implements Serializable
      */
     public function serialize(): string
     {
-        $serialized = json_encode([
+        return \Safe\json_encode([
             'channel' => $this->channel,
             'topic' => $this->topic,
             'query' => serialize(
                 AST::toArray($this->query)
             ),
-            'operation_name' => $this->operationName,
+            'field_name' => $this->fieldName,
             'args' => $this->args,
             'context' => $this->contextSerializer()->serialize($this->context),
         ]);
-
-        // TODO use \Safe\json_encode
-        if (json_last_error() !== JSON_ERROR_NONE) {
-            throw new \Exception('Tried to encode invalid JSON while serializing subscriber data: '.json_last_error_msg());
-        }
-        /** @var string $serialized */
-
-        return $serialized;
     }
 
     /**