+++ tests/Integration/Schema/NodeInterfaceTest.php
@@ -36,7 +36,7 @@ class NodeInterfaceTest extends DBTestCase
 
     public function testCanResolveNodes(): void
     {
-        $this->schema .= '
+        $this->schema .= /** @lang GraphQL */ '
         type User @node(resolver: "Tests\\\Integration\\\Schema\\\NodeInterfaceTest@resolveNode") {
             name: String!
         }
@@ -45,7 +45,7 @@ class NodeInterfaceTest extends DBTestCase
         $firstGlobalId = $this->globalIdResolver->encode('User', $this->testTuples[1]['id']);
         $secondGlobalId = $this->globalIdResolver->encode('User', $this->testTuples[2]['id']);
 
-        $this->graphQL('
+        $this->graphQL(/** @lang GraphQL */ '
         {
             first: node(id: "'.$firstGlobalId.'") {
                 id
@@ -87,7 +87,7 @@ class NodeInterfaceTest extends DBTestCase
      */
     public function testCanResolveModelsNodes(string $directiveDefinition): void
     {
-        $this->schema .= "
+        $this->schema .= /** @lang GraphQL */ "
         type User $directiveDefinition {
             name: String!
         }
@@ -98,7 +98,7 @@ class NodeInterfaceTest extends DBTestCase
         ]);
         $globalId = $this->globalIdResolver->encode('User', $user->getKey());
 
-        $this->graphQL('
+        $this->graphQL(/** @lang GraphQL */ '
         {
             node(id: "'.$globalId.'") {
                 id
@@ -123,10 +123,6 @@ class NodeInterfaceTest extends DBTestCase
     public function modelNodeDirectiveStyles(): array
     {
         return [
-            /*
-             * @deprecated @model will be repurposed in v5
-             */
-            ['@model'],
             ['@node'],
             ['@node(model: "User")'],
         ];