+++ tests/SerializingArrayStore.php
@@ -26,7 +26,7 @@ class SerializingArrayStore extends ArrayStore
     public function get($key)
     {
         if (! isset($this->storage[$key])) {
-            return;
+            return null;
         }
 
         $item = $this->storage[$key];
@@ -36,7 +36,7 @@ class SerializingArrayStore extends ArrayStore
         if ($expiresAt !== 0 && $this->currentTime() > $expiresAt) {
             $this->forget($key);
 
-            return;
+            return null;
         }
 
         return unserialize($item['value']);